requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-huggingface
chromadb
youtube-transcript-api
sentence-transformers
python-dotenv
pytube
